move parser to a lib, and c generator to the outputter
with this commit the .json file gets compiled into some runtime tree,
which is then transformed again into c. Right now there is only .c
support. However, you could also have .cs or any other binded language
in there. Additionally, you should also be able to create the UI
described in .json in runtime. Which might makes this interesting to a
ui designer.

Additionlly, there is now a validator file, which is meant to validate
the whole tree, it is far from perfect yet.
Parser details can be found in the parser, details for the outputter API
in the outputter file.
Details for the tree itself are kept in ui_tree.c